Moving TflApi and PostCodeApi functions to their own Api classes
