+++ BusBoard.ConsoleApp/BusBoard.ConsoleApp.csproj
@@ -64,10 +64,12 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="BusPredictions.cs" />
+    <Compile Include="PostCodeApi.cs" />
     <Compile Include="StopPoints.cs" />
     <Compile Include="Coordinates.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="TflApi.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="App.config" />
