+++ BusBoard.ConsoleApp/BusPredictions.cs
@@ -6,7 +6,7 @@
         public string LineName { get; set; }
         public int TimeToStation { get; set; }
         public string DestinationName { get; set; }
-        
+
         public override string ToString()
         {
             return $"StationName: {StationName}, LineName: {LineName}, TimeToStation: {TimeToStation}, DestinationName: {DestinationName}";
